review: move apis implementations to separate files
